Removed protobuf code and replaced it with script that saves solutions to .yaml files.

+++ irlco/data/generate_training_data.py
@@ -1,8 +1,11 @@
 import copt
 import yaml
+from typing import List, Dict
+from tqdm import tqdm
+import os.path
 
 
-def top_solutions(problem_instance: ProblemInstance, n_top: int) -> List[Dict]:
+def top_solutions(problem_instance, n_top: int) -> List[Dict]:
     """
     Finds the n_top best solutions using brute force search.
     Args:
@@ -15,17 +18,18 @@ def top_solutions(problem_instance: ProblemInstance, n_top: int) -> List[Dict]:
     """
     # Brute force all possible solutions - bruteForce returns a list ordered by measure
     all_solutions = copt.bruteForce(problem_instance)
+    # Get best n_top solutions
     best_solns = all_solutions[:n_top]
     # If any of the best solutions are failures, return nothing
-    for i in range(n_top):
+    for i in range(len(best_solns)):
         if best_solns[i]['success'] == 0:
             return []
     # Otherwise, return the best solutions
     return all_solutions[:n_top]
 
 
-if __name__ == '__main__':
-    config_file = open("generator_config.yaml")
+def generate_data(config_path):
+    config_file = open(config_path)
     configs = yaml.load_all(config_file)
 
     for config in configs:
@@ -34,9 +38,22 @@ if __name__ == '__main__':
         nb_top_solutions = config['nb_top_solutions']
         output_file = config['output_file']
 
-        out = open(output_file, 'w')
+        # Check if file exists already - if it does, don't generate it again
+        if os.path.exists(output_file):
+            continue
+        else:
+            out = open(output_file, 'w')
 
-        for i in range(nb_instances):
+        for i in tqdm(range(nb_instances)):
             problem = copt.getProblem(instance_size)
             best_solns = top_solutions(problem, nb_top_solutions)
+            # Get rid of path data since we don't need it, and add instance description to each dict
+            for soln in best_solns:
+                del soln['pathData']
+                soln['instance'] = problem
             yaml.dump_all(best_solns, out)
+
+
+if __name__ == '__main__':
+    generate_data("irl_data_config.yaml")
+    generate_data("test_data_config.yaml")
